+++ ghost/core/core/server/data/seeders/importers/StripeProductsImporter.js
