+++ ghost/core/package.json
@@ -60,6 +60,7 @@
   },
   "dependencies": {
     "@extractus/oembed-extractor": "3.2.1",
+    "@faker-js/faker": "7.6.0",
     "@isaacs/ttlcache": "1.4.1",
     "@sentry/node": "7.120.3",
     "@slack/webhook": "7.0.5",
@@ -72,7 +73,6 @@
     "@tryghost/constants": "0.0.0",
     "@tryghost/custom-fonts": "0.0.0",
     "@tryghost/custom-theme-settings-service": "0.0.0",
-    "@tryghost/data-generator": "0.0.0",
     "@tryghost/database-info": "0.3.27",
     "@tryghost/debug": "0.1.32",
     "@tryghost/domain-events": "0.0.0",
@@ -121,6 +121,7 @@
     "@tryghost/promise": "0.3.12",
     "@tryghost/recommendations": "0.0.0",
     "@tryghost/request": "1.0.8",
+    "@tryghost/root-utils": "0.3.30",
     "@tryghost/security": "0.0.0",
     "@tryghost/social-urls": "0.1.43",
     "@tryghost/string": "0.2.12",
@@ -149,6 +150,7 @@
     "cookie-session": "2.1.0",
     "cookies": "0.9.1",
     "cors": "2.8.5",
+    "csv-writer": "1.6.0",
     "date-fns": "2.30.0",
     "dompurify": "3.2.3",
     "downsize": "0.0.8",
@@ -207,6 +209,7 @@
     "otplib": "12.0.1",
     "papaparse": "5.3.2",
     "path-match": "1.2.4",
+    "probability-distributions": "0.9.1",
     "probe-image-size": "7.2.3",
     "rss": "1.2.2",
     "sanitize-html": "2.14.0",
