--- ghost/data-generator/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-    plugins: ['ghost'],
-    extends: [
-        'plugin:ghost/node'
-    ],
-    rules: {
-        'ghost/filenames/match-exported-class': [1, null, true]
-    }
-};
