--- ghost/data-generator/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@tryghost/data-generator",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/data-generator",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "@tryghost/debug": "0.1.32",
-    "c8": "8.0.1",
-    "knex": "2.4.2",
-    "mocha": "10.8.2",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@faker-js/faker": "7.6.0",
-    "@tryghost/root-utils": "0.3.30",
-    "@tryghost/string": "0.2.12",
-    "csv-writer": "1.6.0",
-    "probability-distributions": "0.9.1"
-  }
-}
